+++ agency/execution/executor/customization_points.hpp
@@ -2,7 +2,6 @@
 
 #include <agency/execution/executor/customization_points/async_execute.hpp>
 #include <agency/execution/executor/customization_points/bulk_async_execute.hpp>
-#include <agency/execution/executor/customization_points/bulk_sync_execute.hpp>
 #include <agency/execution/executor/customization_points/bulk_then_execute.hpp>
 #include <agency/execution/executor/customization_points/future_cast.hpp>
 #include <agency/execution/executor/customization_points/make_ready_future.hpp>