+++ agency/execution/executor/detail/utility.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
+#include <agency/execution/executor/detail/utility/blocking_bulk_twoway_execute.hpp>
 #include <agency/execution/executor/detail/utility/bulk_async_execute_with_one_shared_parameter.hpp>
 #include <agency/execution/executor/detail/utility/bulk_async_execute_without_shared_parameters.hpp>
 #include <agency/execution/executor/detail/utility/bulk_async_execute_with_void_result.hpp>