+++ agency/execution/executor/detail/utility/blocking_bulk_twoway_execute.hpp
@@ -9,6 +9,8 @@
 
 namespace agency
 {
+namespace detail
+{
 
 
 __agency_exec_check_disable__
@@ -18,12 +20,13 @@ template<class E, class Function, class ResultFactory, class... Factories,
         >
 __AGENCY_ANNOTATION
 detail::result_of_t<ResultFactory()>
-bulk_sync_execute(const E& exec, Function f, executor_shape_t<E> shape, ResultFactory result_factory, Factories... shared_factories)
+blocking_bulk_twoway_execute(const E& exec, Function f, executor_shape_t<E> shape, ResultFactory result_factory, Factories... shared_factories)
 {
   detail::always_blocking_executor<E> blocking_exec(exec);
   return agency::bulk_async_execute(blocking_exec, f, shape, result_factory, shared_factories...).get();
 }
 
 
+} // end detail
 } // end agency